quant-system-hf/pyproject.toml
[project]
name = "quant-system-hf"
version = "0.2.0"
description = "Hedge-fund style systematic trading system scaffold"
requires-python = ">=3.10"
dependencies = [
  "pandas>=2.0",
  "numpy>=1.24",
  "pyyaml>=6.0",
  "duckdb>=0.10",
  "pyarrow>=14.0",
  "fastapi>=0.110",
  "uvicorn>=0.27",
  "streamlit>=1.30",
  "scikit-learn>=1.3",
  "statsmodels>=0.14",
  "python-dateutil>=2.9",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

[tool.ruff]
line-length = 100
